Arm Utilizer
This program focus on the calculations of final transition matrix by
evaluating the forward kinematics parameters given by user. User can
enter as many joints as they want to the linked list structure and have
the ability to update or delete any joints they want. User Interface is
simple as needed and easy to understand for all kind of users.